tvforms: do not avoid padding in TDataRec
The reason for this is that TDescrRec mimics TMemoData, which is part of the library API and uses ushort as length type. So, replacing ushort with uint does not solve the issue but rather breaks the expectations of TMemo::getData/setData. Because the padding is located at the end of the struct, getData/setData still work.

+++ examples/tvforms/genparts.h
@@ -58,10 +58,9 @@ const int
     qtyWidth       =   6,
     descrLen       = 512;            // Length of text array
 
-
-struct TDescrRec 
+struct TDescrRec // See TMemoData.
 {
-    uint textLen;
+    ushort textLen;
     char textData[descrLen];
 };
 
@@ -100,8 +99,8 @@ TDataRec data[dataCount] =
 
 void initDescrLengths()
 {
-  for ( int i = 0 ; i < dataCount ; i++ )
-      data[i].descr.textLen = strlen(data[i].descr.textData);
+    for ( int i = 0 ; i < dataCount ; i++ )
+        data[i].descr.textLen = strlen(data[i].descr.textData);
 }
 
 TForm *makeForm()
@@ -175,7 +174,5 @@ TForm *makeForm()
 
     return f;
 }
-
-
 
 #endif  // __GENPARTS_H
